First Homework: knit a markdown file
Complete all **Exercises**, and submit answers to **Questions** on the
Coursera
platform.
</div>

The goal of this lab is to introduce you to
R and RStudio, which you'll be using
throughout the course both to learn
the statistical concepts discussed in the
course and to analyze real
data and come to informed conclusions. To straighten
out which is
which: R is the name of the programming language itself and RStudio
is
a convenient interface.

As the labs progress, you are encouraged to
explore beyond what the labs dictate;
a willingness to experiment will
make you a much better programmer. Before we
get to that stage,
however, you need to build some basic fluency in R. Today we
begin with
the fundamental building blocks of R and RStudio: the interface,

reading in data, and basic commands.